+++ cli/run.ts
@@ -1,10 +1,4 @@
-import lume from "../mod.ts";
-import { toFileUrl } from "../deps/path.ts";
-import { isUrl } from "../core/utils/path.ts";
-import { getConfigFile } from "../core/utils/lume_config.ts";
-import { log } from "../core/utils/log.ts";
-
-import type Site from "../core/site.ts";
+import { createSite } from "./utils.ts";
 
 /** Run one or more custom scripts */
 export async function run(
@@ -21,33 +15,4 @@ export async function run(
       break;
     }
   }
-}
-
-/** Create a site instance */
-export async function createSite(config?: string): Promise<Site> {
-  let url: string | undefined;
-
-  if (config && isUrl(config)) {
-    url = config;
-  } else {
-    const path = await getConfigFile(config);
-
-    if (path) {
-      url = toFileUrl(path).href;
-    }
-  }
-
-  if (url) {
-    log.info(`Loading config file <gray>${url}</gray>`);
-    const mod = await import(url);
-    if (!mod.default) {
-      log.critical(
-        `[Lume] Missing Site instance! Ensure your config file does export the Site instance as default.`,
-      );
-      throw new Error("Site instance is not found");
-    }
-    return mod.default;
-  }
-
-  return lume();
 }