+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [Unreleased]
 ### Added
 - New function `site.parseFilename`, to register new custom parsers to extract data from filenames.
+- Reload the build after change the `_config.ts` file.
 - New plugin `sri`.
 - Improved plugin docs with links to the online documentation.
 - New functions `nav.nextPage()` and `nav.previousPage()` for `nav` plugin.